feat(ui): extract theme to lib/app-theme.ts and add SessionProvider to providers.tsx
